no double vars
Former-commit-id: 51e90c0bcc84b809b8750c5fec3eb0d2d438434d [formerly ff215681431641380fa86ee98984508eb08fe219]
Former-commit-id: 331650911f84c1def3b9c4a72536e5c1d92f5874
Former-commit-id: 2079c3712bbbbfaae45160b8685a7f01be41e479
Former-commit-id: 740f409e5e9e44a4c399794d0ce21f9af5e81a6a

+++ finalMVA/mkNtuple/makeTuple.C
@@ -503,18 +503,18 @@ Bool_t makeTuple::Process(Long64_t entry)
   int nbjets_m = 0; 
 
   TLorentzVector p4met;
-  double met = *MET;
-  double met_phi = *MET_phi;
-  double apt = TMath::Abs(met);
-  double met_x =  apt*TMath::Cos(met_phi);
-  double met_y =  apt*TMath::Sin(met_phi);
+  float met = *MET;
+  float met_phi = *MET_phi;
+  float apt = TMath::Abs(met);
+  float met_x =  apt*TMath::Cos(met_phi);
+  float met_y =  apt*TMath::Sin(met_phi);
   p4met.SetPxPyPzE( met_x, met_y, 0, met);
 
   TLorentzVector lepton;
   lepton.SetPtEtaPhiE(*lepton_pt, *lepton_eta, *lepton_phi, *lepton_e);
 
-  double transverseM = transverseMass(lepton, p4met);
-  double lepDphi = lepton.DeltaPhi(p4met);
+  float transverseM = transverseMass(lepton, p4met);
+  float lepDphi = lepton.DeltaPhi(p4met);
 
   //Selection Option
   bool isQCD = transverseM < 10 && met < 10 && lepDphi < 1;
@@ -1023,13 +1023,13 @@ void makeTuple::Terminate()
 
 }
 
-double makeTuple::transverseMass( const TLorentzVector & lepton, const TLorentzVector & met){
+float makeTuple::transverseMass( const TLorentzVector & lepton, const TLorentzVector & met){
 
   TLorentzVector leptonT(lepton.Px(),lepton.Py(),0.,lepton.E()*TMath::Sin(lepton.Theta()));
   TLorentzVector metT(met.Px(), met.Py(), 0, met.E());
 
   TLorentzVector sumT=leptonT+metT;
-  double out = TMath::Sqrt( sumT.M2() );
+  float out = TMath::Sqrt( sumT.M2() );
 
   return out;
